Reorganize repository - source files to top level
