+++ rollup.config.js
@@ -1,7 +1,7 @@
 import buble from 'rollup-plugin-buble';
 
 export default {
-  input: 'src/index.js',
+  input: 'topologica.js',
   output: {
     format: 'umd',
     name: 'Topologica',
